+++ common/src/main/java/com/predator/common/registry/init/item/block/PredatorSpawnEggItems.java
@@ -10,28 +10,17 @@ import net.minecraft.world.entity.Mob;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.SpawnEggItem;
 
-import java.util.function.Supplier;
-
 public class PredatorSpawnEggItems {
 
     public static final BLibRegistry<Item> REGISTRY = Predator.MOD.registries().create(BuiltInRegistries.ITEM);
 
-    public static final BLibHolder<SpawnEggItem> YAUTJA_SPAWN_EGG = create(
-        "yautja",
-        PredatorEntityTypes.YAUTJA,
-        0xB9A86C,
-        0x5A4728
-    );
+    public static final BLibHolder<SpawnEggItem> YAUTJA_SPAWN_EGG = create("yautja", PredatorEntityTypes.YAUTJA);
 
-    private static <E extends Mob> BLibHolder<SpawnEggItem> create(
-        String baseId,
-        Supplier<EntityType<E>> entityTypeSupplier,
-        int primaryColor,
-        int secondaryColor
-    ) {
-        var supplier = Predator.MOD.factories()
-            .createSpawnEggSupplier(entityTypeSupplier, primaryColor, secondaryColor, new Item.Properties());
-        return REGISTRY.createHolder(baseId + "_spawn_egg", supplier);
+    private static <E extends Mob> BLibHolder<SpawnEggItem> create(String path, BLibHolder<EntityType<E>> holder) {
+        return REGISTRY.createHolder(
+            path + "_spawn_egg",
+            Predator.MOD.factories().createSpawnEggSupplier(holder, 0xFFFFFF, 0xFFFFFF, new Item.Properties())
+        );
     }
 
     public static void initialize() {